Create script `nixos-fde-config`
nixos-fde-config is a script for configuring a NixOS system with full disk
encryption.

This script creates a LUKS2 encrypted container on the given main disk,
then adds three ext4-formatted LVM volumes (root, swap, home).

The LUKS2 container's header-file, along with a decryption key-file are
added as a compressed, passphrase-protected archive to a UEFI boot partition
on the USB boot key.

After a successful setup, the boot key is required to boot the system.
You will also need to provide your passphrase to mount the filesystem.